refactor: Scope field name to custom field
